fix dtype handling in ConvSwiGLU forward method

+++ models/layers.py
@@ -268,7 +268,7 @@ class ConvSwiGLU(nn.Module):
     def forward(self, x: torch.Tensor, timer: Optional[object] = None, prefix: str = ""):
         gate, up = self.gate_up_proj(x).chunk(2, dim=-1)
         x_ffn = self.act(gate) * up
-        x_conv = self.dwconv(x_ffn.transpose(1, 2).to(x_ffn.dtype))
+        x_conv = self.dwconv(x_ffn.transpose(1, 2).to(self.dwconv.weight.dtype))
         x_conv = x_conv[..., :up.size(1)]
         x_conv = self.act(x_conv)
         x_conv = x_conv.transpose(1, 2).contiguous()
